dns: handle exception in do_sendv for udp
a std::system_error in udp.channel.send(...) can go unhandled, if the
future is not available immediately. For example, ECONNABORTED can go
uncaught and be shown in the log as an Error.
This commit adds a handler to log it as a warning, since the protocol
already handles the case of lost udp packets